+++ tests/testthat/test-add_nevent.R
@@ -95,3 +95,21 @@ test_that("add_nevent error with bad inputs", {
     NULL
   )
 })
+
+# add_nevent.tbl_surfit --------------------------------------------------------
+
+test_that("add_nevent.tbl_surfit", {
+  library(survival)
+
+  tbl_survfit <-
+    list(
+      survfit(Surv(ttdeath, death) ~ 1, trial),
+      survfit(Surv(ttdeath, death) ~ trt, trial)
+    ) %>%
+    tbl_survfit(times = c(12, 24))
+
+  expect_error(
+    add_nevent(tbl_survfit),
+    NA
+  )
+})